use size_t for arguments to iconv, not int. Fixes 64 bit platforms.
2000-05-24  Matt Wilson  <msw@redhat.com>

     * modules/basic/basic.c (conv_8bit): use size_t for arguments to
     iconv, not int.  Fixes 64 bit platforms.

     * modules/basic/basic.c (conv_euc): likewise

+++ modules/basic/basic-x.c
@@ -270,9 +270,9 @@ conv_8bit (CharCache  *cache,
   const char *p;
   
   const char *inptr = input;
-  int inbytesleft;
+  size_t inbytesleft;
   char *outptr = &outbuf;
-  int outbytesleft = 1;
+  size_t outbytesleft = 1;
 
   _pango_utf8_iterate (input, &p, NULL);
   inbytesleft = p - input;
@@ -294,9 +294,9 @@ conv_euc (CharCache  *cache,
   const char *p;
 
   const char *inptr = input;
-  int inbytesleft;
+  size_t inbytesleft;
   char *outptr = outbuf;
-  int outbytesleft = 2;
+  size_t outbytesleft = 2;
 
   _pango_utf8_iterate (input, &p, NULL);
   inbytesleft = p - input;
